requirements.txt
-r submodule/HiCDiffusion/requirements.txt
fire
dataclasses-json
transformers==4.46.3
torchtyping
requests
scikit-learn
